Move run loop into libstratis
